+++ Macmonitor/MacMonitor-Bridging-Header.h
@@ -2,6 +2,7 @@
 #define MacMonitor_Bridging_Header_h
 
 #include "SMC.h"
+#include "ProcessIO.h"
 #import "IOReportWrapper.h"
 
 #endif